Move replace_current_process to yash_env crate
The `exec` built-in utility no longer depends on `yash-semantics` crate
since the `replace_current_process` function is now imported from the
`yash-env` crate.

This is part of the effort to decouple `yash-builtin` from
`yash-semantics` (<https://github.com/magicant/yash-rs/issues/625>).
